requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
aiohttp>=3.9.0
markdownify>=0.11.0
nltk>=3.8.0
python-dateutil>=2.8.0
urllib3>=2.1.0
webdriver-manager>=4.0.0 